+++ types/collections_test.go
@@ -5,6 +5,8 @@ import (
 	"time"
 
 	"cosmossdk.io/collections/colltest"
+	"github.com/stretchr/testify/require"
+	"pgregory.net/rapid"
 )
 
 func TestCollectionsCorrectness(t *testing.T) {
@@ -26,5 +28,16 @@ func TestCollectionsCorrectness(t *testing.T) {
 
 	t.Run("Time", func(t *testing.T) {
 		colltest.TestKeyCodec(t, TimeKey, time.Time{})
+	})
+}
+
+func TestLEUint64Key(t *testing.T) {
+	t.Run("conformance", rapid.MakeCheck(func(r *rapid.T) {
+		colltest.TestKeyCodec(t, LEUint64Key, rapid.Uint64().Draw(r, "uint64"))
+	}))
+
+	t.Run("buffer too small", func(t *testing.T) {
+		_, _, err := LEUint64Key.Decode([]byte{0})
+		require.ErrorContains(t, err, "invalid buffer size")
 	})
 }
